pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["dltoolkit", "dltoolkit.*"]

[project]
name = "dltoolkit"
version = "0.1.0"
description = "A convenient toolbox for deep learning"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=1.10.1",
    "datasets>=3.5.0",
    "deepspeed>=0.17.5",
    "einops>=0.8.1",
    "fvcore>=0.1.5.post20221221",
    "hydra-core>=1.3.2",
    "modelscope>=1.29.1",
    "notebook>=7.4.5",
    "opencv-python>=4.11.0.86",
    "pandas>=2.2.3",
    "tensorboard>=2.20.0",
    "timm>=1.0.15",
    "torch>=2.6.0",
    "torchaudio>=2.6.0",
    "torchdata>=0.11.0",
    "torchvision>=0.21.0",
    "transformers",
]


[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple/"

[tool.uv.sources]
torch = [
  { index = "pytorch-cu124", marker = "platform_system != 'Darwin'" },
]
torchvision = [
  { index = "pytorch-cu124", marker = "platform_system != 'Darwin'" },
]

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

